requirements.txt
absl-py==2.1.0
accelerate==0.26.1
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
ansicon==1.89.0
anyio==4.2.0
async-timeout==4.0.3
attrs==23.2.0
bitsandbytes==0.40.2
bleach==6.1.0
blessed==1.20.0
blis==0.7.11
cachetools==5.3.2
catalogue==2.0.10
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
commonmark==0.9.1
confection==0.1.4
cymem==2.0.8
Cython==3.0.8
dataclasses-json==0.6.3
datasets==2.16.1
decorator==5.1.1
dill==0.3.7
et-xmlfile==1.1.0
exceptiongroup==1.2.0
faiss-cpu==1.7.4
fastapi==0.109.0
filelock==3.13.1
frozenlist==1.4.1
fsspec==2023.10.0
getdaft==0.1.20
google-auth==2.27.0
google-auth-oauthlib==1.2.0
gpustat==1.1.1
greenlet==3.0.3
grpcio==1.60.0
h11==0.14.0
h5py==3.10.0
html5lib==1.1
httptools==0.6.1
huggingface-hub==0.20.3
idna==3.6
imagecodecs==2024.1.1
importlib-metadata==7.0.1
Jinja2==3.1.3
jinxed==1.2.1
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.6.2
kaggle==1.6.3
langcodes==3.3.0
langsmith==0.0.83
lightning-utilities==0.10.1
loguru==0.7.2
loralib==0.1.2
ludwig==0.9.3
lxml==5.1.0
Markdown==3.5.2
MarkupSafe==2.1.4
marshmallow==3.20.2
marshmallow-dataclass==8.5.4
marshmallow-jsonschema==0.13.0
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
murmurhash==1.0.10
mypy-extensions==1.0.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.3
nvidia-ml-py==12.535.133
oauthlib==3.2.2
openpyxl==3.1.2
packaging==23.2
pandas==2.1.4
peft==0.7.1
Pillow==9.3.0
preshed==3.0.9
protobuf==4.23.4
psutil==5.9.4
py==1.11.0
py-cpuinfo==9.0.0
pyarrow==15.0.0
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==1.10.14
pydantic_core==2.14.6
Pygments==2.17.2
pyrsistent==0.20.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-slugify==8.0.2
pytz==2023.3.post1
pyxlsb==1.0.10
PyYAML==6.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
retry==0.9.2
rich==12.4.4
rsa==4.9
sacremoses==0.1.1
safetensors==0.4.2
scikit-learn==1.4.0
scipy==1.12.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.25
srsly==2.4.8
starlette==0.35.1
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
text-unidecode==1.3
thinc==8.2.2
threadpoolctl==3.2.0
tifffile==2023.12.9
tokenizers==0.15.1
torchdata==0.7.1
torchinfo==1.8.0
torchtext==0.16.2
tqdm==4.66.1
transformers==4.37.1
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==1.26.18
uvicorn==0.27.0
wasabi==1.1.2
watchfiles==0.21.0
wcwidth==0.2.13
weasel==0.3.4
webencodings==0.5.1
websockets==12.0
Werkzeug==3.0.1
win32-setctime==1.1.0
xlrd==2.0.1
XlsxWriter==3.1.9
xlwt==1.3.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.17.0
